+++ crates/astria-sequencer/src/service/mempool/tests.rs
@@ -34,7 +34,8 @@ async fn future_nonces_are_accepted() {
     let mempool = fixture.mempool();
     let metrics = fixture.metrics();
     let rsp =
-        super::handle_check_tx_request(req, storage.latest_snapshot(), &mempool, metrics).await;
+        super::handle_check_tx_request(req, storage.latest_snapshot(), mempool.clone(), metrics)
+            .await;
     assert_eq!(rsp.code, Code::Ok, "{rsp:#?}");
 
     // mempool should contain single transaction still
@@ -57,7 +58,8 @@ async fn rechecks_pass() {
     let mempool = fixture.mempool();
     let metrics = fixture.metrics();
     let rsp =
-        super::handle_check_tx_request(req, storage.latest_snapshot(), &mempool, metrics).await;
+        super::handle_check_tx_request(req, storage.latest_snapshot(), mempool.clone(), metrics)
+            .await;
     assert_eq!(rsp.code, Code::Ok, "{rsp:#?}");
 
     // recheck also passes
@@ -66,7 +68,8 @@ async fn rechecks_pass() {
         kind: CheckTxKind::Recheck,
     };
     let rsp =
-        super::handle_check_tx_request(req, storage.latest_snapshot(), &mempool, metrics).await;
+        super::handle_check_tx_request(req, storage.latest_snapshot(), mempool.clone(), metrics)
+            .await;
     assert_eq!(rsp.code, Code::Ok, "{rsp:#?}");
 
     // mempool should contain single transaction still
@@ -90,7 +93,8 @@ async fn can_reinsert_after_recheck_fail() {
     let mempool = fixture.mempool();
     let metrics = fixture.metrics();
     let rsp =
-        super::handle_check_tx_request(req, storage.latest_snapshot(), &mempool, metrics).await;
+        super::handle_check_tx_request(req, storage.latest_snapshot(), mempool.clone(), metrics)
+            .await;
     assert_eq!(rsp.code, Code::Ok, "{rsp:#?}");
 
     // remove the transaction from the mempool to make recheck fail
@@ -104,7 +108,8 @@ async fn can_reinsert_after_recheck_fail() {
         kind: CheckTxKind::Recheck,
     };
     let rsp =
-        super::handle_check_tx_request(req, storage.latest_snapshot(), &mempool, metrics).await;
+        super::handle_check_tx_request(req, storage.latest_snapshot(), mempool.clone(), metrics)
+            .await;
     assert_eq!(rsp.code, Code::Err(NonZeroU32::new(9).unwrap()), "{rsp:#?}");
 
     // can re-insert the transaction after first recheck fail
@@ -113,7 +118,8 @@ async fn can_reinsert_after_recheck_fail() {
         kind: CheckTxKind::New,
     };
     let rsp =
-        super::handle_check_tx_request(req, storage.latest_snapshot(), &mempool, metrics).await;
+        super::handle_check_tx_request(req, storage.latest_snapshot(), mempool.clone(), metrics)
+            .await;
     assert_eq!(rsp.code, Code::Ok, "{rsp:#?}");
 }
 
@@ -136,7 +142,8 @@ async fn recheck_adds_non_tracked_tx() {
     let mempool = fixture.mempool();
     let metrics = fixture.metrics();
     let rsp =
-        super::handle_check_tx_request(req, storage.latest_snapshot(), &mempool, metrics).await;
+        super::handle_check_tx_request(req, storage.latest_snapshot(), mempool.clone(), metrics)
+            .await;
     assert_eq!(rsp.code, Code::Ok, "{rsp:#?}");
 
     // mempool should contain single transaction still